Added support for corner ledges
Player is extended with a RayCast2D that checks for corners.

The Corner scene is a scene with an Area2D and two collision boxes that
represent the corners.

The implementation is via a script attached to the TileMap, which checks if it
is an actual corner and draws a Corner node (based on the corner scene) to the
global position of the tile.